Refactoring transaction component
