+++ src/components/Transaction/CardTransaction/TransactionIcon/TransactionIcon.styles.ts
